web-animations/animation-player/animation-player-pause.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player pause() method test</title>
<meta name="assert" content="Suspends the playback of this player by running the procedure to pause a player for this object">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationplayer-pause">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
var TIMEOUT = 50;

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    var times = [-100, 0, 100, ANIMATION_DURATION, ANIMATION_DURATION + 100];

    player.ready.then(t.step_func(function() {
        times.forEach(function(time) {
            player.play();
            player.currentTime = time;
            player.pause();
            assert_equals(player.currentTime, time, 'Wrong current time for time ' + time);
            assert_equals(player.startTime, null, 'Start time should be unresolved ' +
                'for time ' + time);
            assert_equals(player.playState, 'paused', 'AnimationPlayer should be ' +
                'paused for time ' + time);
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that AnimationPlayer.pause() sets start time to unresolved and pauses the player');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = -1;
    var times = [-100, 0, 100, ANIMATION_DURATION, ANIMATION_DURATION + 100];

    player.ready.then(t.step_func(function() {
        times.forEach(function(time) {
            player.play();
            player.currentTime = time;
            player.pause();
            assert_equals(player.currentTime, time, 'Wrong current time for time ' + time);
            assert_equals(player.startTime, null, 'Start time should be unresolved ' +
                'for time ' + time);
            assert_equals(player.playState, 'paused', 'AnimationPlayer should be ' +
                'paused for time ' + time);
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that AnimationPlayer.pause() sets start time to unresolved and pauses the player ' +
    'when playbackRate < 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));

    player.ready.then(t.step_func(function() {
        // wait for 50 ms
        setTimeout(t.step_func(function () {
            player.pause();
            // wait for 50 ms
            setTimeout(t.step_func(function() {
                assert_approx_equals(player.currentTime, TIMEOUT, TIME_EPSILON,
                    'Current time should be paused');
                assert_equals(player.startTime, null, 'Start time should be unresolved');
                assert_equals(player.playState, 'paused', 'AnimationPlayer should be paused');                
                t.done();
            }), TIMEOUT);
        }), TIMEOUT);        
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that AnimationPlayer.pause() call stops current time');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = -1;
    player.currentTime = ANIMATION_DURATION / 2;

    player.ready.then(t.step_func(function() {
        // wait for 50 ms
        setTimeout(t.step_func(function () {
            player.pause();
            // wait for 50 ms
            setTimeout(t.step_func(function() {
                assert_approx_equals(player.currentTime, ANIMATION_DURATION / 2 - TIMEOUT,
                    TIME_EPSILON, 'Current time should be paused');
                assert_equals(player.startTime, null, 'Start time should be unresolved');
                assert_equals(player.playState, 'paused', 'AnimationPlayer should be paused');                
                t.done();
            }), TIMEOUT);
        }), TIMEOUT);        
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that AnimationPlayer.pause() call stops current time when playback rate < 0');
</script>
</body>
